+++ CHANGELOG.md
@@ -8,10 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Geometry hash field populated by event listener
 - Calculation object with hash registry
 - Calculation hash table populated by event listener
-- QCIO ProgramInput generation from Calculation object
+- Interconversion of Calculation and QCIO ProgramInput objects
 
 ### Fixed
-- Specify Calculation->Energy and Geometry->Energy as many-to-one relationships
+- Specify Calculation-Energy and Geometry-Energy as many-to-one relationships
 
 ## [0.0.2] - 2026-01-28
 ### Added